Give caller access to longopt name as it was parsed
Add 'optname' field to clopts_control, storing a pointer to the longopt
name as it exists in the argument vector, thus an abbreviated longopt
name can be retreived without need for the caller to check the argument
vector. This can be useful for caller error/diagnostic messages to
address the longopt by the same name that was entered by the user.